% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stacked_bar.R
\name{plot_stacked_bar}
\alias{plot_stacked_bar}
\title{plot_stacked_bar}
\usage{
plot_stacked_bar(data, x_col, stack_col, fill_col)
}
\arguments{
\item{data}{a dataframe}

\item{x_col}{charcater, column name for x group}

\item{stack_col}{charcater, column name for stack group}

\item{fill_col}{charcater, column name for fill group}
}
\value{
a ggplot object
}
\description{
plot stacked bar
}
\examples{
cell_data <- data.frame( Sample = rep(c("Sample1", "Sample2", "Sample3"), each = 4),Type = rep(c("Type1", "Type2", "Type3", "Type4"), 3),Count = c(10, 15, 8, 6, 5, 7, 12, 9, 3, 6, 4, 11))
}
